+++ RELEASES.md
@@ -77,6 +77,7 @@ Upcoming
 - Add writeSuffix parameter to FileHandler (next to writeSuffix attribute)
 - Add file type bin for mime type application/octet-stream in FileViewerServlet
 - Make it possible to generate a WSDL based on WsdlXmlValidator
+- Add Last Message to Show ConfigurationStatus